+++ resources/views/admin/rackcatalogitem/index.blade.php
@@ -1,14 +1,9 @@
 <x-admin-layout>
 
     <div class="flex w-full pl-3 pt-3 gap-3">
-        <div class="w-64 rounded-xl border border-gray-200 bg-white shadow-xs p-4 dark:bg-gray-800 dark:border-gray-700">
-            <div class="h-8 text-cerulean-500 text-center font-CoconPro">
-                {{ __('Katalogelemente Gesamt') }}
-            </div>
-            <div class="h-10 text-chathams-blue-800 dark:text-gray-100 text-center font-CoconPro text-4xl">
-                {{ $rackCatalogItemsCount }}
-            </div>
-        </div>
+        <x-adminkachel :label="__('Katalogelemente Gesamt')">
+            {{ $rackCatalogItemsCount }}
+        </x-adminkachel>
     </div>
 
     <x-sitetopmenu />