+++ resources/views/admin/role/index.blade.php
@@ -1,23 +1,13 @@
 <x-admin-layout>
 
     <div class="flex w-full pl-3 pt-3 gap-3">
-        <div class="w-64 rounded-xl border border-gray-200 bg-white shadow-xs p-4 dark:bg-gray-800 dark:border-gray-700">
-            <div class="h-8 text-cerulean-500 text-center font-CoconPro">
-                {{ __('Rollen Gesamt') }}
-            </div>
-            <div class="h-10 text-chathams-blue-800 dark:text-gray-100 text-center font-CoconPro text-4xl">
-                {{ $roleCount }}
-            </div>
-        </div>
+        <x-adminkachel :label="__('Rollen Gesamt')">
+            {{ $roleCount }}
+        </x-adminkachel>
 
-        <div class="w-64 rounded-xl border border-gray-200 bg-white shadow-xs p-4 dark:bg-gray-800 dark:border-gray-700">
-            <div class="h-8 text-cerulean-500 text-center font-CoconPro">
-                {{ __('Zuletzt hinzugefügt') }}
-            </div>
-            <div class="h-10 text-chathams-blue-800 dark:text-gray-100 text-center font-CoconPro text-2xl">
-                {{ $roleLastAdded->name }}
-            </div>
-        </div>
+        <x-adminkachel :label="__('Zuletzt hinzugefügt')" art="name">
+            {{ $roleLastAdded->name }}
+        </x-adminkachel>